migration

+++ backend/CatchEmAll.DataAccess/Providers/Migrator.cs
@@ -1,4 +1,5 @@
 using CatchEmAll.Options;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
 using System.Threading.Tasks;
 
@@ -19,16 +20,22 @@ namespace CatchEmAll.Providers
     {
       try
       {
-        await this.context.Database.EnsureDeletedAsync();
-        await this.context.Database.EnsureCreatedAsync();
-        //await this.context.Database.MigrateAsync();
+        if (this.options.RecreateDatabaseOnStartup)
+        {
+          await this.context.Database.EnsureDeletedAsync();
+          await this.context.Database.EnsureCreatedAsync();
+        }
+        else
+        {
+          await this.context.Database.MigrateAsync();
+        }
       }
       catch
       {
         if (this.options.DeleteDatabaseOnMigrationFailure)
         {
-          //await context.Database.EnsureDeletedAsync();
-          //await context.Database.MigrateAsync();
+          await context.Database.EnsureDeletedAsync();
+          await context.Database.MigrateAsync();
         }
         else
         {
